+++ manufacturing.c
@@ -12,7 +12,18 @@
 #include "state.h"			// For the event handlers
 #include "statemodel.h" // For the other states
 #include <stdio.h>
+#include <pthread.h>
 
+extern int order_size;
+
+typedef struct {
+	int capacity, 
+			duration;
+} args_t;
+
+args_t * x;
+
+//***************never declare a variable in a header or include a .c in header of .c file
 // Create an object of the manufacturing state
 state_t manufacturing = {
 		default_event_handler,  // order_received
@@ -25,9 +36,6 @@ state_t manufacturing = {
 		entry_to_manufacturing, // entry_to
 		exit_from_manufacturing,// exit_from
 };
-
-extern int order_size;
-extern int args_t;
 
 // Call the manufacture completed handler, exit from,
 // return address of shipping
@@ -46,23 +54,26 @@ state_t* manufacture_failed()
 	updateStats(FAIL);
 	return &accepting;
 }
-
+// need a sema
 // Set the action when entering the manufacturing state
 // Create five threads to represent manufacturing lines
 static void entry_to_manufacturing()
 {
+	// need a sema here create mutex set to init sem (x=1) ****
 	//dispatchFactoryLines();
-	pthread_t thread;
-
+	pthread_t threads;
 	int i, lines;
 	lines = 5;
-
+  
+	// create thread for each man line
 	for(i = 0; i < lines; i++)
 	{
 	 	x = malloc(sizeof (args_t));
-		pthread_create(&thread, NULL, dispatchFactoryLines, (void *) x);
-	free(x);
+	  x->capacity = random() % (50 + 1 - 10) + 10;
+	  x->duration = random() % (5 + 1 - 1) + 1;
+		pthread_create(&threads, NULL, dispatchFactoryLines, (void *) x);
 	printf("Order Size = %d\n", order_size);
+	}
 }
 
 // Set the action when exiting the manufacturing state
@@ -71,4 +82,49 @@ static void exit_from_manufacturing()
 	shutDownFactoryLines();
 }
 	
-		
+void * dispatchFactoryLines(void * p) {
+	printf("Dispatching Factory Lines\n");
+	args_t * x = (args_t *) p;
+	free(x);
+	int c = x->capacity;
+	int d = x->duration;
+}
+
+// each thread will look at how many items reaning to be made it end else it continues
+// while loop while(more2make() > 0 {
+// more2make function returns how mnay you will make, either your capacity or whatever is left, get capactiy as agr, and check how many are left to make
+// decrement number of items left to be made 
+// 2bmade, myCap, myTask <- before 
+// after return 2Bmade to zero is no tasks left
+// ^^^^ more to make function needs to be protected by a mutex
+// mutex inside loop or not inside the more2makefunc? grab mutex closer to critical point, need var iter and made for while loop
+// sleep (duration);
+// ++ iters;
+// made += y;
+// }
+// print made / iterations   ... i think this is happeining in dispatch factory lines
+// pthread_exit( retval/made/iters) pass a return value on exit (look at lab that does that)
+/*
+shutdownfl()
+{
+ join 5 threads (in order)
+ destroy semaphores
+ free pointer that had return
+*/
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
